Air Quality Sensor Implemented
Implemented the Air Quality Sensor, It now works, but it's not as accurate as intended.

Apparently there is a load resistor of only up to 1k so I had to downgrade the status to basically range the accuracy of this module instead. So now I am dealing with the status of this module.