Added WarmUp Period
Added a warm up period for the Air Quality sensor, and changed up the values to make the range make more sense. There is still a bug for changing status but I tried to resolve it. Yet to no avail. I'll keep it as it is and focus on the compass for now.