Updated memory
Worked on the memory side of the project, Was finally able to bring it down from 74 to 63.3 percent, it's now back up since I'll be focusing on the Compass now. But it is less from what it was before thankfully. I did things dynamically instead when it comes to different screen states. Now starting tomorrow I will be working on the compass.