Updated the Framework
I used up all of the memory on the Arduino Uno. So I'll need to migrate immediately to the ESP32. I'll do my best to find ways in reducing memory later on. So hopefully maybe next time I can still use the Arduino. Need to test the button mechanics though.